🏗 vue-dash: Move layout components to dedicated folders
