+++ packages/docs/src/data/contentExplorerItems.ts
@@ -114,33 +114,56 @@ export const contentExplorerItems: TreeviewItem[] = [
 						name: 'layout',
 						children: [
 							{
-								name: 'tests',
+								name: 'AppFooter',
 								children: [
 									{
-										name: 'AppFooter.spec.ts',
-										file: 'ts'
+										name: 'tests',
+										children: [
+											{
+												name: 'AppFooter.spec.ts',
+												file: 'ts'
+											}
+										]
 									},
 									{
-										name: 'AppHeader.spec.ts',
-										file: 'ts'
+										name: 'AppFooter.vue',
+										file: 'vue'
 									},
 									{
-										name: 'HeaderMenu.spec.ts',
+										name: 'index.ts',
 										file: 'ts'
 									}
 								]
 							},
 							{
-								name: 'AppFooter.vue',
-								file: 'vue'
-							},
-							{
-								name: 'AppHeader.vue',
-								file: 'vue'
-							},
-							{
-								name: 'HeaderMenu.vue',
-								file: 'vue'
+								name: 'AppHeader',
+								children: [
+									{
+										name: 'tests',
+										children: [
+											{
+												name: 'AppHeader.spec.ts',
+												file: 'ts'
+											},
+											{
+												name: 'HeaderMenu.spec.ts',
+												file: 'ts'
+											}
+										]
+									},
+									{
+										name: 'AppHeader.vue',
+										file: 'vue'
+									},
+									{
+										name: 'HeaderMenu.vue',
+										file: 'vue'
+									},
+									{
+										name: 'index.ts',
+										file: 'ts'
+									}
+								]
 							},
 							{
 								name: 'index.ts',
