+++ packages/vue-cli-plugin-vue-dash/generator/template/src/components/layout/AppFooter/AppFooter.vue
