+++ packages/vue-cli-plugin-vue-dash/generator/template/src/components/layout/AppFooter/index.ts
@@ -1,0 +1,3 @@
+import AppFooter from './AppFooter.vue';
+
+export default AppFooter;
