+++ packages/vue-cli-plugin-vue-dash/generator/template/src/components/layout/AppFooter/tests/AppFooter.spec.ts
@@ -4,15 +4,15 @@ import { Wrapper } from '@vue/test-utils';
 import { mountComponent } from '@/tests-unit';
 import { html } from '@cnamts/vue-dot/tests/utils/html';
 
-import AppHeader from '../AppHeader.vue';
+import AppFooter from '../';
 
 let wrapper: Wrapper<Vue>;
 
 // Tests
-describe('AppHeader', () => {
+describe('AppFooter', () => {
 	it('renders correctly', () => {
 		// Mount component
-		wrapper = mountComponent(AppHeader);
+		wrapper = mountComponent(AppFooter);
 
 		expect(html(wrapper)).toMatchSnapshot();
 	});
