+++ packages/vue-cli-plugin-vue-dash/generator/template/src/components/layout/AppHeader/index.ts
@@ -1,0 +1,3 @@
+import AppHeader from './AppHeader.vue';
+
+export default AppHeader;
